Tests: Use fully qualified paths to the current R and Rscript
PATH may not be set to point to the current R version, or even any R
version at all.  This was revealed (although very obtusely, and only
with much digging) via failures during rhub's automated testing on
Fedora with R-devel.  It makes sense though, and is the equivalent of
using $^X in Perl.